requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pandas==2.2.2
SQLAlchemy==2.0.34
psycopg[binary,pool]==3.2.9
alembic==1.13.2
python-dateutil==2.9.0.post0
tzdata>=2024.1
numpy==1.26.4
websockets==12.0
pytz==2024.1
